refactor: isolate shared MCP error contracts
Move the mapped error envelope and terms classifier out of the code-navigation mapper so service error mapping does not load code-navigation taxonomy. Preserve existing mapper re-exports and terms remediation payloads for current callers.
