+++ packages/mcp/src/shared/githits-service-error-map.ts
@@ -3,8 +3,8 @@ import {
   AuthenticationError,
   FetchTimeoutError,
 } from "@githits/core-internal";
-import type { MappedError } from "./code-navigation-error-map.js";
-import { mapTermsAcceptanceError } from "./code-navigation-error-map.js";
+import type { MappedError } from "./mapped-error.js";
+import { mapTermsAcceptanceError } from "./terms-acceptance-error-map.js";
 
 /**
  * Classify errors from the GitHits API service into the shared envelope used
